Recreate revision with autogenerate enum
