+++ backend/src/alembic/versions/2026_07_02_1333-3db6d67a5f48_add_tool_call_and_tool_result_values_to_.py
@@ -1,0 +1,45 @@
+"""Add tool_call and tool_result values to chatrole enum
+
+Revision ID: 3db6d67a5f48
+Revises: 9252bbb0fae2
+Create Date: 2026-07-02 13:33:56.097364
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+import sqlmodel
+from alembic_postgresql_enum import TableReference
+
+# revision identifiers, used by Alembic.
+revision: str = '3db6d67a5f48'
+down_revision: Union[str, Sequence[str], None] = '9252bbb0fae2'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.sync_enum_values(
+        enum_schema='public',
+        enum_name='chatrole',
+        new_values=['user', 'assistant', 'tool_call', 'tool_result'],
+        affected_columns=[TableReference(table_schema='public', table_name='chat_messages', column_name='role')],
+        enum_values_to_rename=[],
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.sync_enum_values(
+        enum_schema='public',
+        enum_name='chatrole',
+        new_values=['user', 'assistant'],
+        affected_columns=[TableReference(table_schema='public', table_name='chat_messages', column_name='role')],
+        enum_values_to_rename=[],
+    )
+    # ### end Alembic commands ###
